feat: working on spinners
