--- src/components/spinner.js
@@ -1,7 +1,0 @@
-import { Spinner } from '@wordpress/components';
-
-function showSpinner() {
-	return <Spinner />;
-}
-
-export default showSpinner;
